refactor(react-generative-ui): share the spec-bounding pre-pass across Slack and Teams
The Slack converter carried a byte-identical copy of the Teams boundSpec
walk, and its inline clamp-reason text matched clampReasonDetail exactly.
Move both, plus the four bounding constants they share, to
src/convert/boundSpec.ts and re-export the constants from each protocol's
constants module so every existing import site is unchanged.

No output change: constants, warning strings, and traversal order are
identical, and both converter suites still cover the depth, children,
budget, and cycle clamps.
