+++ .changeset/shy-buses-shave.md
@@ -1,0 +1,5 @@
+---
+"@assistant-ui/react-generative-ui": patch
+---
+
+refactor: share one spec-bounding pre-pass between the Slack and Teams converters
