+++ packages/react-generative-ui/src/slack/constants.ts
@@ -11,34 +11,12 @@ import type {
   SlackTextObject,
 } from "./types";
 
-/** The maximum traversal depth, counted in visited values, not tree levels. */
-export const MAX_TRAVERSAL_DEPTH = 64;
-
-/**
- * The element-nesting ceiling a caller observes. `boundNode` visits an element
- * and its `children` array separately, so one level of nesting spends two of
- * {@link MAX_TRAVERSAL_DEPTH}; it runs before the conversion walk and is never
- * the looser of the two. A root array costs one more level than a root object.
- */
-export const MAX_ELEMENT_DEPTH = MAX_TRAVERSAL_DEPTH / 2;
-
-/**
- * The maximum number of entries kept at any single level (root, or a
- * `children` array at any depth) of the raw spec, before it ever reaches
- * `normalizeSpec`. Bounds a hostile array's reported `length` so the pre-pass
- * can never be made to walk further than this regardless of what the array
- * claims about itself.
- */
-export const CHILDREN_CAP = 200;
-
-/**
- * The total number of nodes the pre-normalization bounding walk will visit
- * across one call, regardless of how many times a shared reference recurs.
- * Bounds the combinatorial work a DAG of shared or self-referential nodes
- * would otherwise force even though each individual array stays within
- * {@link CHILDREN_CAP}.
- */
-export const NODE_BUDGET = 5000;
+export {
+  CHILDREN_CAP,
+  MAX_ELEMENT_DEPTH,
+  MAX_TRAVERSAL_DEPTH,
+  NODE_BUDGET,
+} from "../convert/boundSpec";
 
 /** The block limit for a Slack message. */
 export const MESSAGE_BLOCK_CAP = 50;
