+++ packages/react-generative-ui/src/teams/constants.ts
@@ -17,47 +17,14 @@ export const ADAPTIVE_CARD_SCHEMA =
 export const ATTACHMENT_CONTENT_TYPE =
   "application/vnd.microsoft.card.adaptive";
 
-/** The maximum traversal depth, counted in visited values, not tree levels. */
-export const MAX_TRAVERSAL_DEPTH = 64;
-
-/**
- * The element-nesting ceiling a caller observes. `boundNode` visits an element
- * and its `children` array separately, so one level of nesting spends two of
- * {@link MAX_TRAVERSAL_DEPTH}; it runs before the conversion walk and is never
- * the looser of the two. A root array costs one more level than a root object.
- */
-export const MAX_ELEMENT_DEPTH = MAX_TRAVERSAL_DEPTH / 2;
-
-/**
- * The maximum number of entries kept at any single level (root, or a
- * `children` array at any depth) of the raw spec, before it ever reaches
- * `normalizeSpec`. Bounds a hostile array's reported `length` so the pre-pass
- * can never be made to walk further than this regardless of what the array
- * claims about itself.
- */
-export const CHILDREN_CAP = 200;
-
-/**
- * The total number of nodes {@link boundSpec} will visit across one walk,
- * regardless of how many times a shared reference recurs. Bounds the
- * combinatorial work a DAG of shared or self-referential nodes would
- * otherwise force even though each individual array stays within
- * {@link CHILDREN_CAP}.
- */
-export const NODE_BUDGET = 5000;
-
-export type ClampReason = "children" | "budget" | "cycle" | "depth";
-
-export function clampReasonDetail(reason: ClampReason): string {
-  if (reason === "budget") {
-    return `the tree was truncated after ${NODE_BUDGET} nodes.`;
-  }
-  if (reason === "cycle") return "a self-referencing node was dropped.";
-  if (reason === "depth") {
-    return `nodes deeper than ${MAX_ELEMENT_DEPTH} levels were dropped.`;
-  }
-  return `children were clamped to ${CHILDREN_CAP} entries.`;
-}
+export {
+  CHILDREN_CAP,
+  clampReasonDetail,
+  MAX_ELEMENT_DEPTH,
+  MAX_TRAVERSAL_DEPTH,
+  NODE_BUDGET,
+} from "../convert/boundSpec";
+export type { ClampReason } from "../convert/boundSpec";
 
 /** The maximum number of attachments produced from a root-level Carousel. */
 export const CAROUSEL_ATTACHMENT_CAP = 10;
